+++ README.md
@@ -1,4 +1,5 @@
 # ecommerce project with React
 
 -Vite
--Vercel free account linked to GitHub project
+-Vercel free account linked to GitHub project https://vercel.com/mauros-projects-48744d58/ecommerce/deployments?filterBranch=dev
+-React is using atomic design system (atoms, molecules, organisms, pages)
